Optimize plb.engine.primitive.primitives struct
