--- plb/engine/primitive/primitives/__init__.py
@@ -1,3 +1,0 @@
-from .robot_fk import RobotsControllers
-from .primitives import Primitives
-from .shapes import Sphere, Capsule, RollingPin, Chopsticks, Cylinder, Box
